# backend-go/docs/swagger.yaml
definitions:
  github_com_lingmirror_backend-go_internal_response.PageResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  github_com_lingmirror_backend-go_internal_response.Result:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_domain_integrations.PublishToOzonInput:
    properties:
      account_id:
        type: integer
      currency_code:
        type: string
      image_url:
        type: string
      price:
        type: number
      product_id:
        type: integer
    required:
    - account_id
    - price
    - product_id
    type: object
...
info:
  contact: {}
paths:
  /agentos:
    get:
      description: Get AgentOS cockpit dashboard overview
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: AgentOS overview
      tags:
      - agentos
  /agentos/autonomy:
    get:
      description: Get the autonomy configuration and current level for all agents
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Agent autonomy status
      tags:
      - agentos
  /agentos/status:
    get:
      description: Get the current runtime status of the AgentOS
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: AgentOS system status
      tags:
      - agentos
  /agentos/work-items:
    get:
      description: Get pending work items from all agents
      parameters:
      - description: Max items (default 50)
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by risk level
        in: query
        name: risk_level
        type: string
      - description: Filter by agent ID
        in: query
        name: agent_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: List work items
      tags:
      - agentos
  /agentos/work-items/{id}:
    get:
      description: Get detail of a specific work item by ID
      parameters:
      - description: Work item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Get work item detail
      tags:
      - agentos
...
  /owner/agent-activity:
    get:
      description: Get recent agent activity timeline
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Agent activity
      tags:
      - owner
  /owner/platform-sync:
    get:
      description: Get sync status of all connected platform accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Platform sync status
      tags:
      - owner
  /owner/risk-summary:
    get:
      description: Get owner cockpit risk summary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Risk summary
      tags:
      - owner
  /owner/suggestions:
    get:
      description: Get AI-generated suggestions for the business owner
      parameters:
      - description: Max items (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Owner suggestions
      tags:
      - owner
  /platform-integrations:
    get:
      consumes:
      - application/json
      description: Get paginated list of platform integration accounts
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Filter by platform
        in: query
        name: platform_id
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.PageResult'
      security:
      - BearerAuth: []
      summary: List integrations
      tags:
      - integrations
  /platform-integrations/{id}/sync:
    post:
      description: Trigger a sync of orders and products from the platform
      parameters:
      - description: Integration account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Sync integration
      tags:
      - integrations
  /platform-integrations/{id}/test:
    post:
      description: Test whether a platform integration account has valid credentials
      parameters:
      - description: Integration account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Test integration connection
      tags:
      - integrations
  /platform-integrations/publish-to-ozon:
    post:
      consumes:
      - application/json
      description: Publish a product listing to the Ozon platform
      parameters:
      - description: Publish request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_integrations.PublishToOzonInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_lingmirror_backend-go_internal_response.Result'
      security:
      - BearerAuth: []
      summary: Publish to Ozon
      tags:
      - integrations
...
swagger: "2.0"